feat(count-card): refactor CountCard stories and add WithMenu story
